==> notebooks/1_HDFS_hive_env.ipynb <==
# Explore the data files in HDFS 

Use ***kinit Username*** to access hdfs 

This whole initialization of the environment should be runned once as the tables used for spark are stored in HDFS and queried with HIVE 

Run once 

echo ${USERNAME}

### This first part is just used as exploration of HDFS to find the files to be used.
You can skip this part if you already know where the data is stored 

# The data to be used
hdfs dfs -ls /data/sbb/part_orc/timetables

hdfs dfs -ls /data/sbb/part_orc/istdaten

hdfs dfs -ls /data/sbb/part_orc/timetables

# Initializing database 

Only run this to recunstruct databse, if the database is already in order avoid this part. 
Run if you have problems with `2_Data_prep`. 
If you already have your database skip this part and go to the next Notebook.
<div class="alert alert-warning">
<strong>Stop here before going on:</strong> 
   Just run this query if you don't have the database yet, as it may erease the tables you already have and reconstruct the database.
</div>

beeline -u "${HIVE_JDBC_URL}" -e "SHOW DATABASES LIKE '${USERNAME}';
                                  SHOW TABLES IN '${USERNAME}';
                                  DROP DATABASE IF EXISTS ${USERNAME} CASCADE;
                                  SHOW DATABASES LIKE '${USERNAME}';"

hdfs dfs -ls "/user/${USERNAME}"

hdfs dfs -rm -r -f -skipTrash "/user/${USERNAME:-nobody}/hive"
hdfs dfs -rm -r -f -skipTrash "/user/${USERNAME:-nobody}/.Trash"

hdfs dfs -mkdir -p                                /user/${USERNAME:-nobody}/hive
hdfs dfs -setfacl    -m group::r-x                /user/${USERNAME:-nobody}
hdfs dfs -setfacl    -m other::---                /user/${USERNAME:-nobody}
hdfs dfs -setfacl    -m default:group::r-x        /user/${USERNAME:-nobody}
hdfs dfs -setfacl    -m default:other::---        /user/${USERNAME:-nobody}
hdfs dfs -setfacl -R -m group::r-x                /user/${USERNAME:-nobody}/hive
hdfs dfs -setfacl -R -m other::---                /user/${USERNAME:-nobody}/hive
hdfs dfs -setfacl -R -m default:group::r-x        /user/${USERNAME:-nobody}/hive
hdfs dfs -setfacl -R -m default:other::---        /user/${USERNAME:-nobody}/hive
hdfs dfs -setfacl    -m user:hive:rwx             /user/${USERNAME:-nobody}/hive
hdfs dfs -setfacl    -m default:user:hive:rwx     /user/${USERNAME:-nobody}/hive

hdfs dfs -ls "/user/${USERNAME}"

beeline -u "${HIVE_JDBC_URL}" --silent=true -e "
CREATE DATABASE IF NOT EXISTS ${USERNAME:-nobody} LOCATION '/user/${USERNAME:-nobody}/hive';

USE ${USERNAME:-nobody};"
