==> notebooks/3_5_delays_preprossessing.ipynb <==
import pandas as pd

df = pd.read_csv("sbb_connections_friday_names.csv")

df.route_short_name.nunique()

df.arrival_station.nunique()

df_delays = pd.read_csv("delaysfin.csv")

df_delays = df_delays.dropna(axis='rows')

unique_values_df1 = set(zip(df_delays['haltestellen_name'], df_delays['arrival']))
unique_values_df2 = set(zip(df['arrival_station'], df['arrival_time']))

common_values = unique_values_df1.intersection(unique_values_df2)
percentage = (len(common_values) / len(unique_values_df2)) * 100
print(f"Percentage of unique values in common: {percentage}%")

df["produkt_id"] = df["route_desc"].replace({'B':'bus', 'S':'zug', 'IR': 'zug', 'IC':'zug','RE':'zug','EC':'zug','T':'tram'})

print(df["produkt_id"].unique())

print(df_delays["produkt_id"].unique())

df_delays = df_delays.rename(columns={'arrival': 'arrival_time', 'haltestellen_name': 'arrival_station'})

merged_del_co = df.merge(df_delays, on=["produkt_id", "arrival_time", "arrival_station"], how = "outer")
merged_del_co

merged_del_co = merged_del_co.dropna(axis='rows', subset= ['route_id'])
merged_del_co 

merged_del_co = merged_del_co.fillna(0)
merged_del_co 

merged_del_co['90_perc_arrival'] = merged_del_co['arrival_time'] + pd.to_timedelta(merged_del_co['90thPercentile'], unit='s')
merged_del_co['90_perc_arrival'] = merged_del_co['90_perc_arrival'].astype(str).str[7:]

merged_del_co['95_perc_arrival'] = merged_del_co['arrival_time'] + pd.to_timedelta(merged_del_co['95thPercentile'], unit='s')
merged_del_co['95_perc_arrival'] = merged_del_co['95_perc_arrival'].astype(str).str[7:]

merged_del_co['85_perc_arrival'] = merged_del_co['arrival_time'] + pd.to_timedelta(merged_del_co['85thPercentile'], unit='s')
merged_del_co['85_perc_arrival'] = merged_del_co['85_perc_arrival'].astype(str).str[7:]

merged_del_co['80_perc_arrival'] = merged_del_co['arrival_time'] + pd.to_timedelta(merged_del_co['80thPercentile'], unit='s')
merged_del_co['80_perc_arrival'] = merged_del_co['80_perc_arrival'].astype(str).str[7:]

merged_del_co['75_perc_arrival'] = merged_del_co['arrival_time'] + pd.to_timedelta(merged_del_co['75thPercentile'], unit='s')
merged_del_co['75_perc_arrival'] = merged_del_co['75_perc_arrival'].astype(str).str[7:]

merged_del_co['70_perc_arrival'] = merged_del_co['arrival_time'] + pd.to_timedelta(merged_del_co['70thPercentile'], unit='s')
merged_del_co['70_perc_arrival'] = merged_del_co['70_perc_arrival'].astype(str).str[7:]

merged_del_co['65_perc_arrival'] = merged_del_co['arrival_time'] + pd.to_timedelta(merged_del_co['65thPercentile'], unit='s')
merged_del_co['65_perc_arrival'] = merged_del_co['65_perc_arrival'].astype(str).str[7:]

merged_del_co['60_perc_arrival'] = merged_del_co['arrival_time'] + pd.to_timedelta(merged_del_co['60thPercentile'], unit='s')
merged_del_co['60_perc_arrival'] = merged_del_co['60_perc_arrival'].astype(str).str[7:]

merged_del_co['55_perc_arrival'] = merged_del_co['arrival_time'] + pd.to_timedelta(merged_del_co['55thPercentile'], unit='s')
merged_del_co['55_perc_arrival'] = merged_del_co['55_perc_arrival'].astype(str).str[7:]

merged_del_co['50_perc_arrival'] = merged_del_co['arrival_time'] + pd.to_timedelta(merged_del_co['50thPercentile'], unit='s')
merged_del_co['50_perc_arrival'] = merged_del_co['50_perc_arrival'].astype(str).str[7:]


merged_del_co

merged_del_co.to_csv("sbb_connections_friday_perc.csv")
